% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covHeavyTail-package.R
\name{covHeavyTail-package}
\alias{covHeavyTail-package}
\title{Package for estimation of mean and covariance matrix under heavy-tailed distributions.}
\description{
\code{covHeavyTail} is a package to estimate the mean and covariance matrix
of multivariate time series under heavy-tailed distributions.
}
\section{Functions}{

\code{\link{momentsTyler}}, \code{\link{momentsCauchy}}, and \code{\link{momentsStudentt}}
}

