% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FA_studentT.R
\name{covTFA}
\alias{covTFA}
\title{Robustly estimate parameters of the multivariate Students' t data with (optional) assumption of factor model structure}
\usage{
covTFA(X, r = ncol(X), max_iter = 100, ptol = 0.001, ftol = Inf,
  initializer = NULL, procedure = FALSE)
}
\arguments{
\item{X}{Data matrix}

\item{r}{Interger indicating number of factor dimension (default is \code{ncol(X)}, so no factor model assumption).}

\item{max_iter}{Interger indicating the maximum iterations of estimation method.}

\item{ptol}{Number (>= 0) indicating the tolerance for parameter changing when judge convergence (default is 1e-3).}

\item{ftol}{Number (>= 0) indicating the tolerance for objective changing when judge convergence (default is 0).
Note: it might be time consuming when use objective changing as a convergence judging criterion, especially when X is high-dimensional.}

\item{initializer}{A list of initial value of parameters for starting method.}

\item{procedure}{A logical value indicating whether to recode the procedure by iterations.}
}
\value{
The estimated parameters as a list.
}
\description{
Robust paramater estimation of the multivariate Students' t data with (optional) assumption of factor model structure
}
\examples{
# examples are not yet ready!

}
\author{
Rui ZHOU and Daniel P. Palomar
}
