% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_estim.R
\name{momentsCauchy}
\alias{momentsCauchy}
\title{Estimation of mean and covariance matrix based on fitting data to the Cauchy distribution.}
\usage{
momentsCauchy(X, verbose = FALSE)
}
\arguments{
\item{X}{Data matrix containing the multivariate time series (each column is one time series).}

\item{verbose}{If \code{TRUE}, info about iterations for convergence will be output.}
}
\value{
A list with the following components:
\item{\code{mu}}{mean estimate}
\item{\code{cov}}{covariance matrix estimate}
\item{\code{obj_value_record}}{convergence of objective value vs iterations}
}
\description{
The function \code{momentsCauchy} implements a maximum likelihood estimation (MLE) of the mean and covariance matrix
to fit the data to a Cauchy distribution (Student-t with nv=1).
}
\examples{
data(heavy_data)
res <- momentsStudentt(heavy_data$X)
norm(res$mu - heavy_data$mu, "2")
norm(colMeans(heavy_data$X) - heavy_data$mu, "2")
norm(res$cov - heavy_data$cov, "F")
norm(cov(heavy_data$X) - heavy_data$cov, "F")
}
\references{
Ying Sun, Prabhu Babu, and Daniel P. Palomar, “Regularized Robust Estimation of Mean and Covariance Matrix Under Heavy-Tailed Distributions,”
IEEE Trans. on Signal Processing, vol. 63, no. 12, pp. 3096-3109, June 2015.
}
\author{
Daniel P. Palomar
}
