% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_CauchyTyler.R
\name{momentsStudentt}
\alias{momentsStudentt}
\title{Estimation of mean and covariance matrix based on fitting data to the Student-t distribution.}
\usage{
momentsStudentt(X, nu = NULL, max_iter = 100, method = "ECM",
  verbose = FALSE)
}
\arguments{
\item{X}{Data matrix containing the multivariate time series (each column is one time series).}

\item{nu}{If passed, then \code{nu} will not be optimized and will be set to that value.}

\item{method}{Choice of method c("ECM", "ECME") for the maximization of \code{nu}.}

\item{verbose}{If \code{TRUE}, info about iterations for convergence will be output.}
}
\value{
A list with the following components:
\item{\code{mu}}{mean estimate}
\item{\code{cov}}{covariance matrix estimate}
\item{\code{nu}}{degrees of freedom}
\item{\code{obj_value_record}}{convergence of objective value vs iterations}
}
\description{
The function \code{momentsStudentt} implements a maximum likelihood estimation (MLE) of the mean, covariance matrix,
and degrees of freedom nu to fit the data to a Student-t distribution (via the ECME algorithm).
}
\examples{
data(heavy_data)
res <- momentsStudentt(heavy_data$X)
norm(res$mu - heavy_data$mu, "2")
norm(colMeans(heavy_data$X) - heavy_data$mu, "2")
norm(res$cov - heavy_data$cov, "F")
norm(cov(heavy_data$X) - heavy_data$cov, "F")
}
\references{
Chuanhai Liu and Donald B. Rubin, “ML estimation of the t-distribution using EM and its extensions, ECM and ECME,”
Statistica Sinica (5), pp. 19-39, 1995.
}
\author{
Daniel P. Palomar and Junyan LIU
}
