% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_estim.R
\name{momentsTyler}
\alias{momentsTyler}
\title{Estimation of mean (with median) and of covariance matrix via Tyler's estimator.}
\usage{
momentsTyler(X, verbose = FALSE)
}
\arguments{
\item{X}{Data matrix containing the multivariate time series (each column is one time series).}

\item{verbose}{If \code{TRUE}, info about iterations for convergence will be output.}
}
\value{
A list with the following components:
\item{\code{mu}}{mean estimate}
\item{\code{cov}}{covariance matrix estimate}
\item{\code{obj_value_record}}{convergence of objective value vs iterations}
}
\description{
The function \code{momentsTyler} first estimates a robust mean with the median and then (after removing
the mean) estimates the covariance matrix based on Tyler's estimator.
}
\examples{
data(heavy_data)
res <- momentsStudentt(heavy_data$X)
norm(res$mu - heavy_data$mu, "2")
norm(colMeans(heavy_data$X) - heavy_data$mu, "2")
norm(res$cov - heavy_data$cov, "F")
norm(cov(heavy_data$X) - heavy_data$cov, "F")
}
\references{
Ying Sun, Prabhu Babu, and Daniel P. Palomar, “Regularized Tyler’s Scatter Estimator: Existence, Uniqueness, and Algorithms,”
IEEE Trans. on Signal Processing, vol. 62, no. 19, pp. 5143-5156, Oct. 2014.
}
\author{
Daniel P. Palomar
}
