% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FA_studentT.R
\name{fit_mvt}
\alias{fit_mvt}
\title{Robustly estimate parameters of the multivariate Students' t data with (optional) assumption of factor model structure}
\usage{
fit_mvt(X, factors = ncol(X), max_iter = 100, ptol = 0.001,
  ftol = Inf, method = "ECM", nu = NULL, initializer = NULL,
  return_iterates = FALSE)
}
\arguments{
\item{X}{Data matrix}

\item{factors}{Interger indicating number of factor dimension (default is \code{ncol(X)}, so no factor model assumption).}

\item{max_iter}{Interger indicating the maximum iterations of estimation method.}

\item{ptol}{Number (>= 0) indicating the tolerance for parameter changing when judge convergence (default is Inf).}

\item{ftol}{Number (>= 0) indicating the tolerance for objective changing when judge convergence (default is Inf).
Note: it might be time consuming when use objective changing as a convergence judging criterion, especially when X is high-dimensional.}

\item{initializer}{A list of initial value of parameters for starting method.}

\item{return_iterates}{A logical value indicating whether to recode the procedure by iterations.}
}
\value{
The estimated parameters as a list.
}
\description{
Robust paramater estimation of the multivariate Students' t data with (optional) assumption of factor model structure
}
\examples{
# examples are not yet ready!

}
\references{
Rui ZHOU, Junyan Liu, Sandeep Kumar, and Daniel P. Palomar, "Robust factor analysis parameter estimation,"
Lecture Notes in Computer Science, 2020. <doi:10.1109/TSP.2015.2452219>
}
\author{
Rui ZHOU and Daniel P. Palomar
}
