% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mvt.R
\name{fit_mvt}
\alias{fit_mvt}
\title{Robustly estimate parameters of the multivariate Students' t data with (optional) assumption of factor model structure}
\usage{
fit_mvt(X, factors = ncol(X), max_iter = 100, ptol = 0.001,
  ftol = Inf, method = "ECM", nu = NULL, nu_target = NULL,
  nu_regcoef = 0, nu_regfun = abs, nu_shell = function(x) x,
  initializer = NULL, return_iterates = FALSE,
  nu_target_first_iter = FALSE)
}
\arguments{
\item{X}{Data matrix}

\item{factors}{Interger indicating number of factor dimension (default is \code{ncol(X)}, so no factor model assumption).}

\item{max_iter}{Interger indicating the maximum iterations of estimation method.}

\item{ptol}{Number (>= 0) indicating the tolerance for parameter changing when judge convergence (default is Inf).}

\item{ftol}{Number (>= 0) indicating the tolerance for objective changing when judge convergence (default is Inf).
Note: it might be time consuming when use objective changing as a convergence judging criterion, especially when X is high-dimensional.}

\item{method}{String indicating the method of estimating nu: "ECM" - by maximizing the Q function w.r.t. nu; "ECME" - by maximizing the L function w.r.t. nu.}

\item{nu}{Number (>= 0), if passed, the estimated nu is fixed to this number.}

\item{nu_target}{Number (>= 2), the regularized target of nu.}

\item{nu_regcoef}{Number (>= 0), the coefficience of nu regularized term.}

\item{initializer}{A list of initial value of parameters for starting method.}

\item{return_iterates}{A logical value indicating whether to recode the procedure by iterations.}

\item{scale_correct}{A logical value indicating whether to correct the nu, so that the scale of covariance matrix.}
}
\value{
The estimated parameters as a list.
}
\description{
Robust paramater estimation of the multivariate Students' t data with (optional) assumption of factor model structure
}
\details{
The nu regularized term is added to the nu sub-problem as \code{nu_regcoef * abs(nu - nu_target)}
}
\examples{
# examples are not yet ready!

}
\references{
Rui ZHOU, Junyan Liu, Sandeep Kumar, and Daniel P. Palomar, "Robust factor analysis parameter estimation,"
Lecture Notes in Computer Science, 2020. <https://arxiv.org/abs/1909.12530>
}
\author{
Rui ZHOU and Daniel P. Palomar
}
